requirements.txt
dspy-ai>=2.6.0
mistralai>=1.0.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pyyaml>=6.0.2